Revert "Add selectable Pool Royale traditional table"
